// src/services/central-api/generated/model/get-all-price-tier-item-output.ts
/* tslint:disable */
/* eslint-disable */
/**
 * PL Cental API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface GetAllPriceTierItemOutput
 */
export interface GetAllPriceTierItemOutput {
    /**
     * 
     * @type {number}
     * @memberof GetAllPriceTierItemOutput
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof GetAllPriceTierItemOutput
     */
    'name'?: string | null;
    /**
     * 
     * @type {number}
     * @memberof GetAllPriceTierItemOutput
     */
    'minDay'?: number | null;
    /**
     * 
     * @type {number}
     * @memberof GetAllPriceTierItemOutput
     */
    'maxDay'?: number | null;
    /**
     * 
     * @type {number}
     * @memberof GetAllPriceTierItemOutput
     */
    'sequence'?: number | null;
    /**
     * 
     * @type {boolean}
     * @memberof GetAllPriceTierItemOutput
     */
    'isActive'?: boolean | null;
    /**
     * 
     * @type {string}
     * @memberof GetAllPriceTierItemOutput
     */
    'createBy'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof GetAllPriceTierItemOutput
     */
    'createTimestamp'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof GetAllPriceTierItemOutput
     */
    'updateBy'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof GetAllPriceTierItemOutput
     */
    'updateTimestamp'?: string | null;
}